Removed special module, moved scoped value classes to the main module
