--- context-propagation-test/build.gradle
@@ -1,10 +1,0 @@
-plugins {
-    id 'idea'
-}
-
-dependencies {
-    api project(":context-propagation")
-
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testImplementation 'org.assertj:assertj-core'
-}
