+++ context-propagation/build.gradle
@@ -3,9 +3,6 @@ plugins {
 }
 
 dependencies {
-	testImplementation(project(":context-propagation-test")) {
-	  	exclude module: 'context-propagation'
-	}
 	testImplementation 'org.junit.jupiter:junit-jupiter'
 	testImplementation 'org.assertj:assertj-core'
 	testImplementation 'org.mockito:mockito-core'
