+++ context-propagation/src/main/java/io/micrometer/context/package-info.java
@@ -31,5 +31,3 @@
 @NonNullFields
 package io.micrometer.context;
 
-import io.micrometer.context.util.annotation.NonNullApi;
-import io.micrometer.context.util.annotation.NonNullFields;
