+++ context-propagation/src/test/java/io/micrometer/context/ScopedValueTest.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.context.test;
+package io.micrometer.context;
 
+import io.micrometer.context.scopedvalue.ScopedValue;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
