+++ context-propagation/src/test/java/io/micrometer/context/TestThreadLocalAccessor.java
@@ -17,8 +17,6 @@ package io.micrometer.context;
 
 import java.util.Objects;
 
-import io.micrometer.context.util.annotation.Nullable;
-
 /**
  * ThreadLocalAccessor for testing purposes with a given key and {@link ThreadLocal}
  * instance.
