+++ context-propagation/src/test/java/io/micrometer/context/scopedvalue/ScopedValueThreadLocalAccessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.context.test;
+package io.micrometer.context.scopedvalue;
 
 import io.micrometer.context.ThreadLocalAccessor;
