+++ settings.gradle
@@ -22,4 +22,3 @@ gradleEnterprise {
 }
 
 include 'context-propagation'
-include 'context-propagation-test'
